unify algorithms in a single list, and improve documentation
Since b9e02e015be61903bbee58e3fd349114fa28e0b4, definition of algorithms was
distributed across two files, causing them to be documented in two separate
groups.

This patch:

- re-unifies all algorithms.
- improves documentation.
- adds doc-links to relevant specifications / reference material.
